Restructured Runtime React App form fileUpload resource to sharedResources attachments.
