+++ src/react-apps/applications/runtime/src/sagas/index.ts
@@ -6,7 +6,6 @@ import FormConfigSagas from '../features/form/config/sagas';
 import FormDataSagas from '../features/form/data/sagas';
 import FormDataModelSagas from '../features/form/datamodell/sagas';
 import FormDynamicsSagas from '../features/form/dynamics/sagas';
-import FormAttachments from '../features/form/fileUpload/sagas';
 import FormLayoutSagas from '../features/form/layout/sagas';
 import FormResourceSagas from '../features/form/resources/sagas';
 import FormRulesSagas from '../features/form/rules/sagas';
@@ -14,6 +13,7 @@ import FormUserSagas from '../features/form/user/sagas';
 import FormValidationSagas from '../features/form/validation/sagas';
 import FormWorkflowSagas from '../features/form/workflow/sagas';
 import LanguageSagas from '../features/languages/sagas';
+import FormAttachments from '../sharedResources/attachments/attachmentsSagas';
 
 function* root(): SagaIterator {
   yield fork(FormConfigSagas);
