+++ src/react-apps/applications/runtime/src/sharedResources/attachments/fetch/fetchAttachmentsActions.ts
@@ -1,7 +1,7 @@
 
 import { Action } from 'redux';
-import { IAttachments } from '../..';
-import * as ActionTypes from '../types';
+import { IAttachments } from '..';
+import * as ActionTypes from '../attachmentsActionTypes';
 
 export interface IFetchAttachmentsActionFulfilled extends Action {
   attachments: IAttachments;
