+++ src/react-apps/applications/runtime/src/sharedResources/attachments/upload/uploadAttachmentsSagas.ts
@@ -1,13 +1,13 @@
 import { SagaIterator } from 'redux-saga';
 import { call, select, takeEvery } from 'redux-saga/effects';
-import { IAltinnWindow, IAttachment } from '../../';
-import { getFileUploadComponentValidations } from '../../../../../components/base/FileUploadComponent';
-import { IRuntimeState } from '../../../../../types';
-import { get, post } from '../../../../../utils/networking';
-import FormValidationsDispatcher from '../../../validation/actions';
-import FormFileUploadDispatcher from '../../actions';
-import * as FileUploadActionsTypes from '../../actions/types';
-import * as uploadActions from '../../actions/upload';
+import { IAltinnWindow, IAttachment } from '..';
+import { getFileUploadComponentValidations } from '../../../components/base/FileUploadComponent';
+import FormValidationsDispatcher from '../../../features/form/validation/actions';
+import { IRuntimeState } from '../../../types';
+import { get, post } from '../../../utils/networking';
+import FormFileUploadDispatcher from '../attachmentsActions';
+import * as FileUploadActionsTypes from '../attachmentsActionTypes';
+import * as uploadActions from './uploadAttachmentsActions';
 
 export function* uploadAttachmentSaga(
   { file, attachmentType, tmpAttachmentId, componentId }: uploadActions.IUploadAttachmentAction): SagaIterator {
