+++ src/react-apps/applications/runtime/src/types/index.ts
@@ -2,13 +2,13 @@ import { IFormConfigState } from '../features/form/config/reducer';
 import { IFormDataState } from '../features/form/data/reducer';
 import { IDataModelState } from '../features/form/datamodell/reducer';
 import { IFormDynamicState } from '../features/form/dynamics';
-import { IFormFileUploadState } from '../features/form/fileUpload/reducer';
 import { ILayoutState } from '../features/form/layout/reducer';
 import { IResourceState } from '../features/form/resources/reducer';
 import { IFormUserState } from '../features/form/user/reducer';
 import { IValidationState } from '../features/form/validation/reducer';
 import { IWorkflowState } from '../features/form/workflow/reducer';
 import { ILanguageState } from '../features/languages/reducer';
+import { IFormFileUploadState } from '../sharedResources/attachments/attachmentsReducer';
 
 export interface IRuntimeState {
   formLayout: ILayoutState;
